chore(readme): rename generateData to generateContent
