+++ packages/readme/src/utils/generateContent.js
